Rename Authorizer to BasicAuthorizer, deprecate Authorizer
